Devices

+++ routes/devices/devices-route.js
@@ -6,7 +6,7 @@ const {
   deviceValidation,
   idValidation,
 } = require('../../middlewares/validation.js');
-const authorized = require('../../middlewares/Authorized.js');
+// const authorized = require('../../middlewares/Authorized.js');
 
 // /api/devices
 
@@ -15,33 +15,35 @@ const authorized = require('../../middlewares/Authorized.js');
  * @route   POST /api/devices/
  * @access  Private, Admin
  */
-router.post(
-  '/',
-  restricted,
-  authorized('admin'),
-  deviceValidation,
-  async (req, res) => {
-    try {
-      const { userID } = await userDB.findByUUID(req.headers.decodedToken.uid);
-      if (userID) {
-        const device = await db.add(req.device, userID);
-        res.status(201).json(device);
-      }
-    } catch ({ message }) {
-      res.status(500).json({ message: 'Unable to create device the device' });
+router.post('/', restricted, deviceValidation, async (req, res) => {
+  try {
+    const foundUser = await userDB.findByUUID(req.headers.decodedToken.uid);
+
+    if (foundUser && foundUser.is_admin) {
+      const device = await db.add(req.device);
+      res.status(201).json(device);
+    } else {
+      res.status(403).json({ message: 'Access denied!' });
     }
+  } catch ({ message }) {
+    res.status(500).json({ message: 'Unable to create device the device' });
   }
-);
+});
 
 /**
  * @desc    Get all devices
  * @route   GET /api/devices
  * @access  Private, Admin
  */
-router.get('/', restricted, authorized('admin'), async (req, res) => {
+router.get('/', restricted, async (req, res) => {
   try {
-    const devices = await db.findAll();
-    res.status(200).json(devices);
+    const foundUser = await userDB.findByUUID(req.headers.decodedToken.uid);
+    if (foundUser && foundUser.is_admin) {
+      const devices = await db.findAll();
+      res.status(200).json(devices);
+    } else {
+      res.status(403).json({ message: 'Access denied!' });
+    }
   } catch ({ message }) {
     res.status(500).json({ message: 'Unable to retrieve devices.' });
   }
@@ -54,9 +56,9 @@ router.get('/', restricted, authorized('admin'), async (req, res) => {
 router.get('/user-devices', restricted, async (req, res) => {
   const userUID = req.headers.decodedToken.uid;
   try {
-    const { userID } = await userDB.findByUUID(userUID);
-    if (userID) {
-      const userDevices = await db.findALLByUserID(userID);
+    const foundUser = await userDB.findByUUID(userUID);
+    if (foundUser) {
+      const userDevices = await db.findALLByUserID(foundUser.id);
       res.status(200).json(userDevices);
     } else {
       res.status(404).json({ message: `Devices not found!` });
@@ -73,9 +75,12 @@ router.get('/user-devices', restricted, async (req, res) => {
 router.get('/user-devices/:id', restricted, async (req, res) => {
   const userUID = req.headers.decodedToken.uid;
   try {
-    const { userID } = await userDB.findByUUID(userUID);
-    if (userID) {
-      const userDevice = await db.findByIdAndUserID(userID, req.params.id);
+    const foundUser = await userDB.findByUUID(userUID);
+    if (foundUser) {
+      const userDevice = await db.findByIdAndUserID(
+        foundUser.id,
+        req.params.id
+      );
       res.status(200).json(userDevice);
     } else {
       res.status(404).json({ message: `Device not found!` });
@@ -88,26 +93,21 @@ router.get('/user-devices/:id', restricted, async (req, res) => {
 /**
  * @desc    Get a single device by id
  * @route   GET /api/devices/:id
- * @access  Private, Admin, User
+ * @access  Private, Admin
  */
-router.get(
-  '/:id',
-  restricted,
-  authorized('admin'),
-  idValidation,
-  async (req, res) => {
-    try {
+router.get('/:id', restricted, idValidation, async (req, res) => {
+  try {
+    const foundUser = await userDB.findByUUID(req.headers.decodedToken.uid);
+    if (foundUser && foundUser.is_admin) {
       const device = await db.findById(req.id);
-      if (device) {
-        res.status(200).json(device);
-      } else {
-        res.status(404).json({ message: `Device not found!` });
-      }
-    } catch ({ message }) {
-      res.status(500).json({ message: 'Unable to retrieve the device.' });
+      res.status(200).json(device);
+    } else {
+      res.status(404).json({ message: `Device not found!` });
     }
+  } catch ({ message }) {
+    res.status(500).json({ message: 'Unable to retrieve the device.' });
   }
-);
+});
 
 /**
  * @desc    Update a single device
@@ -117,17 +117,16 @@ router.get(
 router.put(
   '/:id',
   restricted,
-  authorized('admin'),
   idValidation,
   deviceValidation,
   async (req, res) => {
     try {
-      const device = await db.findById(req.id);
-      if (!device) {
-        res.status(404).json({ message: 'Device not found.' });
+      const foundUser = await userDB.findByUUID(req.headers.decodedToken.uid);
+      if (foundUser && foundUser.is_admin) {
+        const updatedDevice = await db.update(req.id, req.update);
+        res.status(201).json(updatedDevice);
       } else {
-        const device = await db.update(req.id, req.update);
-        res.status(201).json(device);
+        res.status(403).json({ message: 'Access denied!' });
       }
     } catch ({ message }) {
       res.status(500).json({ message: 'Unable to update the device.' });
@@ -140,20 +139,22 @@ router.put(
  * @route   DELETE /api/devices/:id
  * @access  Private, Admin
  */
-router.delete(
-  '/:id',
-  restricted,
-  authorized('admin'),
-  idValidation,
-  async (req, res) => {
-    try {
+router.delete('/:id', restricted, idValidation, async (req, res) => {
+  try {
+    const foundUser = await userDB.findByUUID(req.headers.decodedToken.uid);
+    if (foundUser && foundUser.is_admin) {
       const device = await db.remove(req.id);
-      if (!device) res.status(404).json({ message: 'Device not found' });
-      else res.status(200).json({ message: 'Device successfully deleted.' });
-    } catch ({ message }) {
-      res.status(500).json({ message: 'Unable to delete the device.' });
+      if (device) {
+        res.status(200).json({ message: 'Device successfully deleted.' });
+      } else {
+        res.status(404).json({ message: 'Device not found' });
+      }
+    } else {
+      res.status(403).json({ message: 'Access denied!' });
     }
+  } catch ({ message }) {
+    res.status(500).json({ message: 'Unable to delete the device.' });
   }
-);
+});
 
 module.exports = router;
